Move stdin_utils tests to dedicated test file
Problem: Tests for stdin_utils.py (expand_stdin_placeholder, resolve_stdin_placeholder)
were placed in agent_utils_test.py, breaking the codebase pattern where each module
has its own _test.py file.
Fix: Created stdin_utils_test.py and moved the relevant tests and _TtyStringIO helper
class from agent_utils_test.py to follow the established naming convention.
